fix(qtui): read keyfiles via platform::read_file (MSVC build + key hygiene)
Both keyfile call sites in unlock_controller.cpp passed
std::filesystem::path::c_str() straight to std::fopen(). On Windows
path::value_type is wchar_t, so MSVC rejected it -- the only two compile
errors in the Qt UI's first-ever MSVC build.

platform::read_file() is the fix rather than a .string().c_str() patch,
because the local code was a hand-rolled copy of it (its own comment even
said "matching SDL app pattern via platform::read_file") that reintroduced
the exact hazard read_file exists to avoid: its comment notes it sizes the
file first and does one allocation + one read specifically so KEY MATERIAL
isn't strewn across freed heap blocks. The replaced loop grew a vector in
4096-byte chunks, leaving a stale copy of the keyfile in every reallocated
block -- against the "wipe keys" invariant in CLAUDE.md. Using read_file
fixes the portability error and that hygiene regression together, and drops
~20 lines of duplication.

Also fixes the uint64_t samples counter printed with %lu in selftest.cpp:
correct on LP64, wrong on Windows where unsigned long is 32-bit (MSVC
C4477). The other %lu uses there are genuinely unsigned long; left alone.

Verified: build clean and all 46 test binaries pass on Linux against Qt 6.9.3.

+++ src/qtui/selftest.cpp
@@ -959,23 +959,23 @@ int runSelftest(const QString& vaultPath)
                 } else if (audioDriver && std::strcmp(audioDriver, "dummy") == 0) {
                     // Dummy driver: MUST have fallback engaged (only if init succeeded)
                     if (fallback) {
-                        fprintf(stdout, "ASSERT (c) Audio state: Dummy driver, fallback ASSERTED (consumed=%lu)... PASS\n", samplesConsumed);
+                        fprintf(stdout, "ASSERT (c) Audio state: Dummy driver, fallback ASSERTED (consumed=%llu)... PASS\n", static_cast<unsigned long long>(samplesConsumed));
                         assertC = true;
                     } else {
-                        fprintf(stdout, "ASSERT (c) Audio state: Dummy driver but fallback NOT engaged (consumed=%lu)... FAIL\n", samplesConsumed);
+                        fprintf(stdout, "ASSERT (c) Audio state: Dummy driver but fallback NOT engaged (consumed=%llu)... FAIL\n", static_cast<unsigned long long>(samplesConsumed));
                         assertC = false;
                     }
                 } else {
                     // Real device: MUST have samples_consumed > 0 OR explicit fallback
                     if (samplesConsumed > 0) {
-                        fprintf(stdout, "ASSERT (c) Audio state: Real device, samples consumed=%lu... PASS\n", samplesConsumed);
+                        fprintf(stdout, "ASSERT (c) Audio state: Real device, samples consumed=%llu... PASS\n", static_cast<unsigned long long>(samplesConsumed));
                         assertC = true;
                     } else if (fallback) {
                         fprintf(stdout, "ASSERT (c) Audio state: No consuming device available, fallback engaged... PASS\n");
                         assertC = true;
                     } else {
-                        fprintf(stdout, "ASSERT (c) Audio state: No consuming device AND no fallback (consumed=%lu, fallback=%d)... FAIL\n",
-                                samplesConsumed, fallback);
+                        fprintf(stdout, "ASSERT (c) Audio state: No consuming device AND no fallback (consumed=%llu, fallback=%d)... FAIL\n",
+                                static_cast<unsigned long long>(samplesConsumed), fallback);
                         assertC = false;
                     }
                 }
